Vista get_turno terminada (falta update y delete)

+++ src/main/java/com/example/proyectointegradororm/service/OdontologoService.java
@@ -1,20 +1,26 @@
 package com.example.proyectointegradororm.service;
 
 import com.example.proyectointegradororm.domain.Odontologo;
+import com.example.proyectointegradororm.domain.Turno;
 import com.example.proyectointegradororm.repository.OdontologoRepository;
+import com.example.proyectointegradororm.repository.TurnoRespository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 @Service
 public class OdontologoService {
     private OdontologoRepository odontologoRepository;
+    protected TurnoRespository turnoRespository;
 
     @Autowired
-    public OdontologoService(OdontologoRepository odontologoRepository) {
+    public OdontologoService(OdontologoRepository odontologoRepository, TurnoRespository turnoRespository) {
         this.odontologoRepository = odontologoRepository;
+        this.turnoRespository = turnoRespository;
+
     }
 
     public Optional<Odontologo> buscarOdontologo(Long id){
@@ -26,7 +32,16 @@ public class OdontologoService {
     }
 
     public Optional<String> eliminarOdontologo(Long id){
-        odontologoRepository.deleteById(id);
+        Odontologo odontologo = odontologoRepository.findById(id).get();
+        List<Turno> turnoList =  turnoRespository.findAll();
+
+        for (Turno turno : turnoList) {
+            if(turno.getOdontologo().getId() == odontologo.getId()){
+                turnoRespository.delete(turno);
+                odontologoRepository.deleteById(id);
+                return Optional.of("El Odontologo ha sido eliminado correctamente y el turno asignado ha sido cancelado");
+            }
+        }
         return Optional.of("El Odontologo ha sido eliminado correctamente");
     }
